refactor cancel alert and mapping
