+++ Potluck/Potluck/ViewControllers/CreateEventControllers/CreateEvent4-Restrictions/MHPCreateEvent4RestrictionsViewController.swift
@@ -21,15 +21,7 @@ class MHPCreateEvent4RestrictionsViewController: UIViewController {
         // Dispose of any resources that can be recreated.
     }
     
-
-    /*
-    // MARK: - Navigation
-
-    // In a storyboard-based application, you will often want to do a little preparation before navigation
-    override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
-        // Get the new view controller using segue.destinationViewController.
-        // Pass the selected object to the new view controller.
+    fileprivate func cancel() {
+        self.presentCancelAlert(view: self)
     }
-    */
-
 }
